fundlens/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "fundlens"
version = "0.1.0"
description = "PE Fund NAV Bridge & Reporting OpenEnv Environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.3",
    "fastmcp>=3.1.1",
    "pydantic>=2.0",
    "fastapi",
    "uvicorn",
    "gradio>=6.0",
    "openai",
    "openpyxl>=3.1",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fundlens*"]
